refactor the tests
