+++ plugins/edu.uoc.som.wapiml.test/src/edu/uoc/som/wapiml/test/simple/TestAllOfWithRef.java
@@ -1,4 +1,4 @@
-package edu.uoc.som.wapiml.test;
+package edu.uoc.som.wapiml.test.simple;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -21,7 +21,7 @@ import com.github.fge.jsonschema.core.exceptions.ProcessingException;
 
 import edu.uoc.som.wapiml.facade.WAPImlFacade;
 
-@DisplayName("Test allOf with Ref")
+@DisplayName("Test allOf with Ref - simple")
 class TestAllOfWithRef {
 
 	private static ResourceSet RES_SET = new ResourceSetImpl();
@@ -38,7 +38,7 @@ class TestAllOfWithRef {
 	void testGenerateAndSaveClassDiagramURI() {
 		
 	        File input = new File("inputs/allOf-with-ref.json");
-	        File output = new File("outputs/allOf-with-ref.uml");
+	        File output = new File("outputs/simple/allOf-with-ref.uml");
 	        
 	        try {
 	        	WAPImlFacade openAPItoUMLFacade = new WAPImlFacade();
